requirements.txt
# 基础依赖
kivy==2.2.1
kivymd==1.1.1
numpy==1.21.6  # 更老但更稳定的版本，兼容性更好
scipy==1.7.3   # 与numpy 1.21兼容的版本
matplotlib==3.5.3  # 最后支持Python 3.7-3.9的版本

# 移除过重的依赖，使用预编译版本
# torch==2.1.0
# scikit-learn==1.3.2

# 轻量级替代
scikit-learn==1.0.2  # 更稳定的版本

# 音频处理
# pyaudio会在编译时被添加，不需要在这里指定

# 图形和UI
kivy_garden.graph==0.4.0
Pillow==9.5.0  # 最后支持Python 3.7-3.9的版本

# Android特定
pyjnius==1.4.0  # 更稳定的版本
# android库会由buildozer提供

# 工具和优化 - 移除不兼容项
# multiprocessing==2.6.2.1 # Python 3 已内置
# asyncio==3.4.3 # Python 3 已内置
# numba==0.57.1  # 编译时间过长

# 测试工具 - 仅开发环境需要
# pytest==7.4.2  # 不需要包含在应用中

# 文档和开发 - 仅开发环境需要
# sphinx==7.2.6  # 文档生成
# pylint==3.0.1  # 代码质量检查

# 额外添加Buildozer需要的依赖
Cython==0.29.33
virtualenv 